Clean up SEL32 sign comparisons
Store the interval timer interrupt level as uint32 to match the setup
API and interrupt-table indexes. Cast the fixed disk and HSDP track-count
comparisons to the existing signed loop index type when marking the last
generated track label.

+++ simulators/SEL32/sel32_disk.c
@@ -2794,7 +2794,7 @@ int disk_label(UNIT *uptr) {
         label[2] = trk & 0xff;                  /* ltkn  trk */
         label[3] = sec & 0xff;                  /* lid   sector ID */
         label[4] = 0x80;                        /* show good sector */
-        if (i == (tot_tracks-1)) {              /* last track? */
+        if (i == ((int)tot_tracks-1)) {         /* last track? */
             label[3] = 0xff;                    /* lid   show as last track label */
             label[4] |= 0x04;                   /* set last track flag */
         }
